Rename : 네트워크 상태별 노출되는 컴포넌트 패키지 변경
